Add files via upload
Basic python script that builds a chat interface with ChatGPT turbo. Upon launching the script, you will be prompted to enter your API key in the terminal. This key is saved in a config file in the same folder the script is located. This scripts will also create two log files. One stores the last conversation, the other is appended to store all the conversations, all with datestamp. Place the migrate file in the same directory as loafco3.py. This script is called during closing the program and will save the chat history.